qa/tasks: moved OSDThrasher from ceph_manager.py to osd_thrash.py
Fixes: https://tracker.ceph.com/issues/41133
